+++ src/class/OutputFormatters/Warning.php
@@ -11,8 +11,6 @@
         /**
          * Constructor
          * @author Art <a.molcanovas@gmail.com>
-         * @uses   AbstractOutputFormatter::__construct()
-         * @uses   \Symfony\Component\Console\Formatter\OutputFormatterStyle::setForeground()
          */
         function __construct() {
             parent::__construct();